refactor:  자동차 (Car) 클래스 객체 분리
- 기존의 자동차 경주 게임에 포함되어 있던 Car 클래스를 domain 폴더에
  별도의 class 파일로 분리
